# Fix optimize color memoization