--- a/pwndbg/lib/memoize.py
+++ b/pwndbg/lib/memoize.py
@@ -70,6 +70,7 @@ class forever(memoize):
     """
     Memoizes forever - for a pwndbg session or until `_reset` is called explicitly.
     """
+
     caches = []  # type: List[forever]
     kind = "forever"