@gsingh93 2023-01-10
I think it'd also be neat if we could collect some stats around how many cache misses were happening so we could catch stuff like this sooner.